src/egl: scale capture region by estimated output fractional scale
wayland doesn't provide per-output fractional scale information,
so we have to calculate the approximate fractional scale ourselves

+++ src/transform.c
@@ -178,7 +178,7 @@ bool region_contains(const region_t * region, const region_t * output) {
     return true;
 }
 
-void region_scale(region_t * region, int32_t scale) {
+void region_scale(region_t * region, float scale) {
     region->x *= scale;
     region->y *= scale;
     region->width *= scale;
